+++ graphql-plugin/src/main/kotlin/com/suparnatural/plugins/graphql/config/SuparnaturalGraphqlExtension.kt
@@ -1,4 +1,4 @@
-package com.suparnatural.plugin.graphql.config
+package com.suparnatural.plugins.graphql.config
 
 object SuparnaturalGraphqlExtension {
     val packageName = "com.suparnatural.output"
