+++ graphql-plugin/src/main/kotlin/com/suparnatural/plugins/graphql/models/Container.kt
@@ -1,4 +1,4 @@
-package com.suparnatural.plugin.graphql.models
+package com.suparnatural.plugins.graphql.models
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
