+++ graphql-plugin/src/main/kotlin/com/suparnatural/plugins/graphql/processors/FragmentProcessor.kt
@@ -1,14 +1,16 @@
-package com.suparnatural.plugin.graphql.processors
+package com.suparnatural.plugins.graphql.processors
 
 import com.squareup.kotlinpoet.*
-import com.suparnatural.plugin.graphql.config.SuparnaturalGraphqlExtension
-import com.suparnatural.plugin.graphql.models.Fragment
+import com.suparnatural.plugins.graphql.config.SuparnaturalGraphqlExtension
+import com.suparnatural.plugins.graphql.models.Fragment
 
 
 fun processFragments(fragments: List<Fragment>, config: SuparnaturalGraphqlExtension): FileSpec {
     val fileSpec = FileSpec.builder(config.packageName, FragmentsContainer)
 
-    val containerInterface = TypeSpec.interfaceBuilder(ClassName(config.packageName, FragmentsContainer))
+    val containerInterface = TypeSpec.interfaceBuilder(ClassName(config.packageName,
+        FragmentsContainer
+    ))
 
     fragments.forEach { fragment ->
         processFieldGroup(fragment, containerInterface, config, true)
