+++ graphql-plugin/src/main/kotlin/com/suparnatural/plugins/graphql/processors/utils.kt
@@ -1,9 +1,9 @@
-package com.suparnatural.plugin.graphql.processors
+package com.suparnatural.plugins.graphql.processors
 
 import com.squareup.kotlinpoet.*
 import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
-import com.suparnatural.plugin.graphql.config.SuparnaturalGraphqlExtension
-import com.suparnatural.plugin.graphql.models.Field
+import com.suparnatural.plugins.graphql.config.SuparnaturalGraphqlExtension
+import com.suparnatural.plugins.graphql.models.Field
 
 val KnownTypes = mutableMapOf(
         "Int" to INT,
@@ -20,7 +20,13 @@ fun propertyTypeName(inputType: String, knownTypes: Map<String, TypeName>, confi
                 .matchEntire(inputType)!!
                 .groups[1]!!
                 .value
-        List::class.asTypeName().parameterizedBy(propertyTypeName(innerType, knownTypes, config))
+        List::class.asTypeName().parameterizedBy(
+            propertyTypeName(
+                innerType,
+                knownTypes,
+                config
+            )
+        )
     } else {
         knownTypes[stripped] ?: ClassName(config?.packageName ?: "", stripped)
     }
@@ -38,7 +44,11 @@ fun strippedType(input: String): String = input.replace(Regex("[\\[\\]?!]"), "")
 fun addPropertiesMutating(fields: List<Field>, interfaceSpec: TypeSpec.Builder?, classSpec: TypeSpec.Builder, config: SuparnaturalGraphqlExtension? = null) {
     val constructorSpec = FunSpec.constructorBuilder()
     fields.forEach {
-        val propertyType = propertyTypeName(it.type, KnownTypes, config)
+        val propertyType = propertyTypeName(
+            it.type,
+            KnownTypes,
+            config
+        )
         interfaceSpec?.addProperty(it.responseName, propertyType)
         constructorSpec.addParameter(it.responseName, propertyType)
         val classPropertySpec = PropertySpec
