+++ graphql-plugin/src/main/resources/META-INF/gradle-plugins/com.suparnatural.plugins.graphql.properties
@@ -1,0 +1,1 @@
+implementation-class=com.suparnatural.plugins.graphql.SuparnaturalGraphQlPlugin
